requirements.txt
discord.py == 1.4.1
aiohttp == 3.6.2 